Application setup simpliciation
Application screen and tile setup moved to class constructors from main taking in input from console to setup things like various window settings like window dimensions and if it's fullscreen etc...

In the future this input will only be taken if a configuration file containing the information is not present. The input will then create the file to save the user choices.